+++ server/src/modes/languageModelCache.ts
@@ -6,8 +6,12 @@ export interface LanguageModelCache<T> {
   dispose(): void;
 }
 
-export function getLanguageModelCache<T>(maxEntries: number, cleanupIntervalTimeInSec: number, parse: (document: TextDocument) => T): LanguageModelCache<T> {
-  let languageModels: { [uri: string]: { version: number, languageId: string, cTime: number, languageModel: T } } = {};
+export function getLanguageModelCache<T>(
+  maxEntries: number,
+  cleanupIntervalTimeInSec: number,
+  parse: (document: TextDocument) => T
+): LanguageModelCache<T> {
+  let languageModels: { [uri: string]: { version: number; languageId: string; cTime: number; languageModel: T } } = {};
   let nModels = 0;
 
   let cleanupInterval: NodeJS.Timer;
@@ -56,7 +60,6 @@ export function getLanguageModelCache<T>(maxEntries: number, cleanupIntervalTime
         }
       }
       return languageModel;
-
     },
     onDocumentRemoved(document: TextDocument) {
       const uri = document.uri;